Fix logging to work more generally
The problem was initialization order.  The loggers were being
created before the configuration was loaded.  This was done
because the initialization was done in __main__, but that's at the end
of the file, and the loggers are defined at the beginning.

The solution is that we need a file which is *only* main, and which
imports everything *after* it configures logging.